IGNITE-27178 Changed PendingRows method type from int to long.

+++ modules/table/src/main/java/org/apache/ignite/internal/table/distributed/StorageUpdateHandler.java
@@ -539,7 +539,7 @@ public class StorageUpdateHandler {
      *
      * @return Total number of pending row IDs.
      */
-    public int getPendingRowCount() {
+    public long getPendingRowCount() {
         return pendingRows.getPendingRowCount();
     }
