+++ modules/table/src/main/java/org/apache/ignite/internal/table/distributed/replicator/PendingRows.java
@@ -93,9 +93,9 @@ public class PendingRows {
      *
      * @return Total number of pending row IDs.
      */
-    public int getPendingRowCount() {
+    public long getPendingRowCount() {
         return txsPendingRowIds.values().stream()
-                .mapToInt(Set::size)
+                .mapToLong(Set::size)
                 .sum();
     }
